+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "hierarchical-prophet"
+name = "prophetverse"
 version = "0.0.1"
 description = ""
 authors = ["Felipe Angelim <felipeangelim@gmail.com>"]
