+++ src/prophetverse/univariate/model.py
